requirements.txt
# AI
anthropic>=0.28.0

# Browser automation
playwright>=1.44.0

# Web dashboard
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
itsdangerous>=2.2.0

# Database
sqlalchemy>=2.0.29

# Config
pydantic-settings>=2.2.1
pyyaml>=6.0.1
python-dotenv>=1.0.0

# CLI
click>=8.1.7
rich>=13.7.1

# PDF generation
weasyprint>=62.3
markdown>=3.6

# Scheduling
apscheduler>=3.10.4

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Utilities
aiofiles>=23.2.1
tenacity>=8.2.3
python-dateutil>=2.9.0
